update webapp v1: llm chat page, knowledge chat page
