--- README.md
@@ -1,3 +1,0 @@
-# llm-chat-webapp
-Web APP template based on streamlit for creating LLM Chat and Langchain APP demo quickly
-
