--- configs/__init__.py
@@ -1,4 +1,0 @@
-from .model_config import *
-from .server_config import *
-
-VERSION = "v1.0"
