+++ webapp-v1/.streamlit/config.toml
@@ -1,0 +1,5 @@
+[theme]
+base="dark"
+
+[browser]
+gatherUsageStats = false
