+++ webapp-v1/configs/__init__.py
@@ -1,0 +1,3 @@
+from .config import *
+
+VERSION = "v1.0"
