+++ webapp-v1/requirements-webapp.txt
@@ -1,0 +1,8 @@
+pydantic
+streamlit>=1.26.0
+streamlit-option-menu>=0.3.6
+streamlit-antd-components>=0.1.11
+streamlit-chatbox>=1.1.6
+streamlit-aggrid>=0.3.4.post3
+httpx~=0.24.1
+watchdog
