--- webui_pages/__init__.py
@@ -1,4 +1,0 @@
-from .dialogue import dialogue_page, chat_box
-from .knowledge_base import knowledge_base_page
-from .model_config import model_config_page
-# from .utils import *
