--- webui_pages/dialogue/__init__.py
@@ -1,1 +1,0 @@
-from .dialogue import dialogue_page, chat_box
