--- webui_pages/knowledge_base/__init__.py
@@ -1,1 +1,0 @@
-from .knowledge_base import knowledge_base_page
