--- webui_pages/knowledge_base_config/__init__.py
@@ -1,1 +1,0 @@
-from .knowledge_base_config import knowledge_base_config_page
