--- webui_pages/model_config/__init__.py
@@ -1,1 +1,0 @@
-from .model_config import model_config_page
